评论操作初步实现，优化喜欢列表去掉原生sql部分

+++ models/favorite.go
@@ -1,7 +1,7 @@
 package models
 
 type Like struct {
-	ID      int `gorm:"primaryKey"`
-	UserId  int `column:"user_id"`
-	VideoId int `column:"video_id"`
+	ID      int64 `gorm:"primaryKey"`
+	UserId  int64 `column:"user_id"`
+	VideoId int64 `column:"video_id"`
 }
